+++ src/UI/Fl_PADnoteOvertonePosition.h
@@ -119,7 +119,7 @@ class PADnoteOvertonePosition: public Fl_Box, public Fl_Osc_Widget
             //normalize
             float max=0;
             for (unsigned i=0; i<nsamples; i++){
-                const float x=fabs(data[i]);
+                const float x=fabsf(data[i]);
                 if (max<x) max=x;
             }
             if (max<0.000001) max=1.0;
